LIKE statement with % and underscore _
I used LIKE Statement with WHERE Clause. and it's different from Where  A = B. 

With =, it will return the output to the EXACT MATCH while the Where A Like '%B_' would return output that similar to B. 

I have broken it down into detail about WHERE A LIKE B use with % and _. Please take a look.